feat: seed mongo data
