+++ API/src/scripts/mongo_data/datas.json
@@ -1,0 +1,32 @@
+[
+  {
+    "data": "Bonjour"
+  },
+  {
+    "data": "Comment ça va ?"
+  },
+  {
+    "data": "Je m'appelle Ylan"
+  },
+  {
+    "data": "J'aime programmer"
+  },
+  {
+    "data": "Le projet annuel est intéressant"
+  },
+  {
+    "data": "J'adore apprendre de nouvelles choses"
+  },
+  {
+    "data": "La technologie est fascinante"
+  },
+  {
+    "data": "J'aime travailler en équipe"
+  },
+  {
+    "data": "Le développement web est passionnant"
+  },
+  {
+    "data": "Je suis impatient de voir les résultats de notre projet"
+  }
+]
