+++ API/src/scripts/mongo_data/messages.json
@@ -1,0 +1,26 @@
+[
+  {
+    "id": "1",
+    "name": "Mathias",
+    "socketID": "abc123",
+    "text": "Bonjour tout le monde !",
+    "createdAt": "2023-10-01T12:00:00Z",
+    "room": "general"
+  },
+  {
+    "id": "2",
+    "name": "Alice",
+    "socketID": "def456",
+    "text": "Salut Mathias !",
+    "createdAt": "2023-10-01T12:05:00Z",
+    "room": "general"
+  },
+  {
+    "id": "3",
+    "name": "Bob",
+    "socketID": "ghi789",
+    "text": null,
+    "createdAt": "2023-10-01T12:10:00Z",
+    "room": "general"
+  }
+]
