+++ API/src/scripts/mongo_data/posts.json
@@ -1,0 +1,138 @@
+
+[
+  {
+    "userId": 1,
+    "neighborhoodId": "quartier1",
+    "content": "Ceci est le contenu du post 1",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:00:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 2,
+    "neighborhoodId": "quartier2",
+    "content": "Ceci est le contenu du post 2",
+    "type": "image",
+    "createdAt": "2023-10-01T12:01:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 3,
+    "neighborhoodId": "quartier3",
+    "content": "Ceci est le contenu du post 3",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:02:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 4,
+    "neighborhoodId": "quartier4",
+    "content": "Ceci est le contenu du post 4",
+    "type": "image",
+    "createdAt": "2023-10-01T12:03:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 5,
+    "neighborhoodId": "quartier5",
+    "content": "Ceci est le contenu du post 5",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:04:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 6,
+    "neighborhoodId": "quartier6",
+    "content": "Ceci est le contenu du post 6",
+    "type": "image",
+    "createdAt": "2023-10-01T12:05:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 7,
+    "neighborhoodId": "quartier7",
+    "content": "Ceci est le contenu du post 7",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:06:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 8,
+    "neighborhoodId": "quartier8",
+    "content": "Ceci est le contenu du post 8",
+    "type": "image",
+    "createdAt": "2023-10-01T12:07:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 9,
+    "neighborhoodId": "quartier9",
+    "content": "Ceci est le contenu du post 9",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:08:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 10,
+    "neighborhoodId": "quartier10",
+    "content": "Ceci est le contenu du post 10",
+    "type": "image",
+    "createdAt": "2023-10-01T12:09:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 11,
+    "neighborhoodId": "quartier11",
+    "content": "Ceci est le contenu du post 11",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:10:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 12,
+    "neighborhoodId": "quartier12",
+    "content": "Ceci est le contenu du post 12",
+    "type": "image",
+    "createdAt": "2023-10-01T12:11:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 13,
+    "neighborhoodId": "quartier13",
+    "content": "Ceci est le contenu du post 13",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:12:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 14,
+    "neighborhoodId": "quartier14",
+    "content": "Ceci est le contenu du post 14",
+    "type": "image",
+    "createdAt": "2023-10-01T12:13:00.000Z",
+    "updatedAt": null,
+    "images": []
+  },
+  {
+    "userId": 15,
+    "neighborhoodId": "quartier15",
+    "content": "Ceci est le contenu du post 15",
+    "type": "texte",
+    "createdAt": "2023-10-01T12:14:00.000Z",
+    "updatedAt": null,
+    "images": []
+  }
+]
